CMI: Fix disappearing notation preview bug
Previously, makeNotationPreviewRange() compared the left side
of a preview rect with the left side of the clipping rect.  This
meant that notes that began to the left of the clipping rect, but
ended inside, would be dropped.  Now the right side of
the preview rect is used and previews no longer disappear
at the left side.

Another use case is resizing a segment from the start.  Previews
would disappear in this case as well.  Now they stay.