Get rid of MidiProgram::operator==()
It wasn't a proper and complete operator==(), so it didn't deserve
the name.  Instead, two partialCompare() functions have been added.
This led to the discovery of two more bugs similar to Bug #1443
but related to the Program and Variation comboboxes in the MIPP.
These have been fixed with this commit.