MidiFile: Improve Instrument setup on import
This fixes bug #1439, MIDI Import Modifies Bank Selects.  Now,
all bank selects and program changes are imported as-is.  Once
a track is imported, the events at time zero are checked to see
if any of them can be moved to the Instrument configuration.

This has caused a cosmetic regression in the names of Segments.
They can end up as "General MIDI Device General MIDI Device #n".
Looking at this next.