Push calls to sendChannelSetup() up
This is the Instrument::setFixedChannel() branch.

The idea here is to move sendChannelSetup() out of setters and
other routines up to a place where it's easy to see the redundant
calls.  This will allow consolidation and reduction of waste.
This process will take some time, and things will likely get worse
before they get better.

This work is related to bug #1480.