Fix crash when loading composition
TrackButtons previously was monitoring documentChanged() and only
doing half of what needed to be done.  This led to crashes when
TrackButtons tried to access the old, freed document.

Since TrackButtons is going away if there is a new document, the
documentChanged() handler has been removed.  This removes the access
to freed memory.

Initial analysis by: Daniel Glöckner