FastVector: Remove unnecessary copy functions
The following functions are removed by this commit:

  - iterator_base copy ctor
  - iterator_base copy assignment operator
  - iterator copy assignment operator
  - reverse_iterator copy assignment operator
  - const_iterator copy assignment operator
  - const_reverse_iterator copy assignment operator

The implementations that were provided were no different from the
ones provided by the compiler.  Since the Rule of Three (now Five)
was being broken in each iterator class, later versions of gcc pick
up on this and issue a warning.

See Patch #65